[onert-micro] Huffman WIP

+++ compiler/luci/pass/src/CircleQuantizer.cpp
@@ -626,8 +626,8 @@ void CircleQuantizer::quantize(loco::Graph *g) const
 
   if (_options->query(Options::Algorithm::QuantizeWeights))
   {
-    static const std::vector<std::string> qw_supported_input_model_dtype{"float32"};
-    static const std::vector<std::string> qw_supported_output_model_dtype{"int4", "int8", "int16"};
+    static const std::vector<std::string> qw_supported_input_model_dtype{"uint8"};
+    static const std::vector<std::string> qw_supported_output_model_dtype{"int4", "uint8", "int16"};
     static const std::vector<std::string> qw_supported_granularity{"channel"};
 
     auto input_model_dtype =
